requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-Classful==0.14.2
google-auth==2.8.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kubernetes==23.6.0
MarkupSafe==2.1.1
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
websocket-client==1.3.3
Werkzeug==2.1.2
zipp==3.8.0
